fixed components layout add styled components
