+++ petbook_fe/components/login/LoginSubmit.tsx
@@ -1,7 +1,10 @@
 import ValidationInput from "@components/common/ValidationInput";
+import navigator from "@lib/modules/navigator";
 
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
+
+import { Container, ButtonBox } from "./styled/styledLoginSubmit";
 
 const BACKEND_BASE_URL = process.env.NEXT_PUBLIC_PY_URL;
 
@@ -11,7 +14,6 @@ export const SocialLogin = () => {
   const [href, setHref] = useState("");
 
   useEffect(() => {
-    console.log(window.location, window.location.href);
     setHref(window.location.href);
   }, []);
 
@@ -20,7 +22,7 @@ export const SocialLogin = () => {
       <a
         className="naver"
         href={`${BACKEND_BASE_URL}/naver/login${
-          router.asPath ? "?redirect_url=" + href : ""
+          router.asPath ? `?redirect_url=${href}` : ""
         }`}
       >
         네이버로 로그인
@@ -28,7 +30,7 @@ export const SocialLogin = () => {
       <a
         className="kakao"
         href={`${BACKEND_BASE_URL}/kakao/login${
-          router.asPath ? "?redirect_url=" + href : ""
+          router.asPath ? `?redirect_url=${href}` : ""
         }`}
       >
         카카오로 로그인
@@ -36,7 +38,7 @@ export const SocialLogin = () => {
       <a
         className="google"
         href={`${BACKEND_BASE_URL}/google/login${
-          router.asPath ? "?redirect_url=" + href : ""
+          router.asPath ? `?redirect_url=${href}` : ""
         }`}
       >
         구글로 로그인
@@ -46,58 +48,49 @@ export const SocialLogin = () => {
 };
 
 export const InduceSign = () => {
-  const router = useRouter();
   return (
-    <>
-      <Container onClick={() => router.push("/register")}>
-        <p>아직 Petbook 계정이 없으신가요?</p>
-        <h3>
-          회원가입을 통해 Petbook 유저들과 <br /> 소통할 수 있어요!
-        </h3>
-      </Container>
-    </>
+    <Container onClick={() => navigator("/register")}>
+      <p>아직 Petbook 계정이 없으신가요?</p>
+      <h3>
+        회원가입을 통해 Petbook 유저들과 <br /> 소통할 수 있어요!
+      </h3>
+    </Container>
   );
 };
 
 export const LoginSubmitButton = () => {
-  useEffect(() => {}, []);
   const onSubmit = () => {
     console.log("onSubmit");
   };
-  return (
-    <>
-      <button onClick={onSubmit}>로그인</button>
-    </>
-  );
+  return <button onClick={onSubmit}>로그인</button>;
 };
 
 export const LoginSubmitForm = () => {
-  useEffect(() => {}, []);
   return (
-    <>
-      <form>
-        <ValidationInput
-          submitType={"login"}
-          axiosValue={"email"}
-          current={"이메일"}
-        />
-        <ValidationInput
-          type={"login"}
-          axiosValue={"password"}
-          current={"비밀번호"}
-        />
-      </form>
-    </>
+    <form>
+      <ValidationInput submitType="login" axiosValue="email" current="이메일" />
+      <ValidationInput
+        submitType="login"
+        axiosValue="password"
+        current="비밀번호"
+      />
+    </form>
   );
 };
 
 export const LoginSubmit = () => {
-  useEffect(() => {}, []);
-  return <></>;
+  return (
+    <>
+      <LoginSubmit.LoginSubmitForm />
+      <LoginSubmit.LoginSubmitButton />
+      <LoginSubmit.SocialLogin />
+      <LoginSubmit.InduceSign />
+    </>
+  );
 };
 
-LoginSubmit.LoginSubmitButton = LoginSubmitButton;
 LoginSubmit.LoginSubmitForm = LoginSubmitForm;
+LoginSubmit.LoginSubmitButton = LoginSubmitButton;
 LoginSubmit.SocialLogin = SocialLogin;
 LoginSubmit.InduceSign = InduceSign;
