+++ petbook_fe/components/register/RegisterPasswordFrom.tsx
@@ -1,28 +1,7 @@
-import styled from "styled-components";
 import ValidationInput from "@components/common/ValidationInput";
-// import PassInputRadio from "@components/register/PassInputRadio";
-
-const PassCheckWrap = styled.ul`
-  display: flex;
-  margin-top: 12px;
-  li {
-    margin-right: 20px;
-    input,
-    progress {
-      accent-color: green;
-    }
-    label {
-      margin-left: 7px;
-      color: #ccc4c2;
-      &.active {
-        color: #00cf46;
-      }
-    }
-  }
-`;
+import { PassCheckWrap } from "./styled/styledRegisterForm";
 
 const PasswordInput = () => {
-  //check의 false ture 로 값 조절
   const PassCheckType = [
     {
       type: "영어 대문자 포함",
@@ -35,13 +14,13 @@ const PasswordInput = () => {
   return (
     <div>
       <ValidationInput
-        submitType={"register"}
-        axiosValue={"password"}
-        current={"비밀번호"}
+        submitType="register"
+        axiosValue="password"
+        current="비밀번호"
       />
       <PassCheckWrap>
         {PassCheckType.map((item: any, index: number) => {
-          const type = item.type;
+          const { type } = item;
           return (
             <li key={index}>
               <input type="radio" id={item.value} checked={item.check} />
