+++ petbook_fe/components/register/styled/styledRegisterForm.tsx
@@ -1,0 +1,88 @@
+import styled from "styled-components";
+
+const RegisterFormWrap = styled.div`
+  width: 100%;
+  max-width: 679px;
+  margin: 0 auto;
+  margin-top: 76px;
+  margin-bottom: 87px;
+  > div {
+    margin-bottom: 26px;
+    &:last-child {
+      margin-bottom: 0;
+    }
+    &.flex {
+      > div {
+        width: 50%;
+        &:first-child {
+          margin-right: 9px;
+        }
+      }
+    }
+  }
+  .box {
+    position: relative;
+    label {
+      position: absolute;
+      top: 50%;
+      left: 30px;
+      opacity: 68%;
+      color: #000000;
+      transform: translateY(-50%);
+    }
+    input {
+      padding-left: 140px;
+      line-height: 30px;
+      &::placeholder {
+        color: #b7b7b7;
+      }
+    }
+  }
+  .err_box {
+    margin-top: 4px;
+    font-size: 15px;
+    font-weight: 300;
+    color: #ff6e4e;
+    line-height: 30px;
+  }
+
+  &.small {
+    max-width: 520px;
+    text-align: center;
+    p {
+      margin-bottom: 14px;
+      font-size: 20px;
+    }
+    h3 {
+      margin-bottom: 55px;
+      font-size: 28px;
+    }
+    .submit_sucess_img {
+      height: 297px;
+      width: 100%;
+      margin-bottom: 28px;
+      background-color: var(--test);
+    }
+  }
+`;
+
+const PassCheckWrap = styled.ul`
+  display: flex;
+  margin-top: 12px;
+  li {
+    margin-right: 20px;
+    input,
+    progress {
+      accent-color: green;
+    }
+    label {
+      margin-left: 7px;
+      color: #ccc4c2;
+      &.active {
+        color: #00cf46;
+      }
+    }
+  }
+`;
+
+export { RegisterFormWrap, PassCheckWrap };
